feat: addmin page component로 분리
