+++ src/refactoring/components/index.ts
@@ -1,3 +1,3 @@
 export * from "./common";
-export * from "./AdminPage";
 export * from "./cart";
+export * from "./admin";
